+++ back/models/Tag.js
@@ -51,6 +51,8 @@ function Tag() {
   this.delete = (tag_id) => {
     logger.debug({ name: `${this.name}.delete()`, data: tag_id }, this.procId, 'method');
 
+    if (!/^[0-9]+$/i.test(tag_id)) return { validationError: 'Invalid ID' };
+
     return db.remove({ id: { field: this.idField, value: tag_id }, table: this.table }, this.procId);
   };
