+++ back/models/Thread.js
@@ -71,6 +71,8 @@ function Thread() {
   this.getByBoard = async (board_id) => {
     logger.debug({ name: `${this.name}.getByBoard()`, data: board_id }, this.procId, 'method');
 
+    if (!/^[0-9]+$/i.test(board_id)) return { validationError: 'Invalid ID' };
+
     let threads = await db.select(
       {
         table: this.table,
@@ -96,6 +98,8 @@ function Thread() {
 
   this.delete = (thread_id) => {
     logger.debug({ name: `${this.name}.delete()`, data: thread_id });
+
+    if (!/^[0-9]+$/i.test(thread_id)) return { validationError: 'Invalid ID' };
 
     return db.remove({ table: this.table, id: { field: this.idField, value: thread_id } }, this.procId);
   };
@@ -127,6 +131,8 @@ function Thread() {
 
   this.getBoardId = async (thread_id) => {
     logger.debug({ name: `${this.name}.getBoardId()`, data: thread_id }, this.procId, 'method');
+
+    if (!/^[0-9]+$/i.test(thread_id)) return { validationError: 'Invalid ID' };
 
     const thread = await db.select(
       { table: this.table, filters: [{ field: this.idField, value: thread_id }] },
